Move some functions from backup to utils
